Flatten the configuration file parser.
Split up the current method into a few smaller ones to hopefully make it easier
to add changes and see how the configuration file works.

The main features are:
+ Each section gets its own parsing helper function and you can easily add
  more.
+ Instead of having a string array -> get a matching index -> switch/case on
  the index, use a std::map and match directly on the input. This reduces the
  need for us to manually count indicies when reading the code and having
  someone insert a property "out of order" won't hurt anything now.
+ There's some duplicated code for getting a list of {strings, doubles, ints}
  which have been rolled up into functions.

I'm not sure this is actually the best split -- it seems like an abuse of
lambdas, but I personally find it more readable than the previous version, and
the functionality should be the same. An alternative (and how I would have
done it in plain C) would be to create a named enum and define a function to
map each string -> enum type, then use a switch case like the existing code
with 0/1/2 replaced with the named enum values.